ASNA Token (#97)
* Create astar_asset_ASNA

* Add files via upload

* Update astar_asset_ASNA

Astarnauts (ASNA) is a community-driven meme token deployed on the Astar EVM Chain. It is designed not only to engage crypto enthusiasts with its fun and creative spirit but also to serve as an experimental project for learning and exploring the capabilities of blockchain technology. The primary objective behind ASNA is to foster a community of like-minded individuals who are eager to gain hands-on experience with tokenomics, blockchain development, and community management.

The Astarnauts project seeks to build a welcoming and educational environment where both newcomers and experienced crypto users can collaborate and share knowledge. By participating in the ASNA ecosystem, supporters have the chance to contribute to a growing community and explore the potential of the Astar network, while also enjoying the light-hearted and entertaining aspects of a meme token.

Our mission includes:

Educational Value: Providing insights and resources for those interested in understanding how tokens are created and managed on the Astar EVM Chain.
Community Engagement: Building a strong and active community, emphasizing collaboration, and offering support for members interested in blockchain technology.
Transparency: We are committed to maintaining open communication about the development and plans of the project, while always welcoming constructive feedback from our supporters.
Join us on this journey as we explore the exciting and ever-evolving world of blockchain technology, and have fun while doing it!

Useful Links:

Website: https://astarnauts.netlify.app/
Twitter: https://x.com/Astarnauts_ASNA
Telegram: https://t.me/Astarnauts

* Update astar_asset_ASNA 2

* Update and rename astar_asset_ASNA to astar_erc20_ASNA

edit name

* Rename astar_asset_ASNA.png to astar_erc20_ASNA.png

edit name

* Rename astar_erc20_ASNA to astar_erc20_ASNA.json

* Update astar_erc20_ASNA.json

* Rename astar_erc20_ASNA.json to astar-erc20-ASNA.json

* Rename astar_erc20_ASNA.png to astar-erc20-ASNA.png

* Update astar-erc20-ASNA.json

* Rename astar-erc20-ASNA.png to astar_erc20_ASNA.png

* Update astar-erc20-ASNA.json