UPDATE WUD logo
Aware of using an account which does not have admin rights, because they are burned. Could this PR be merged by hand?

I was also wondering why the Coingecko price is not visable, I dubble checked the coingecko ID, this could be correct. I remember it was working on Hydration and never on the AH part on Subscan. But lately both are not showing the price