UPDATE WUD logo (#141)
* Add WUD info

Sorry, the owner and admin account are burned. Could you please add the files without the signature?

* Update hydration-custom-WUD.json

* Add WUD to assethub

* Add WUD Coingecko API ID to hydration

* UPDATE WUD logo

Aware of using an account which does not have admin rights, because they are burned. Could this PR be merged by hand?

I was also wondering why the Coingecko price is not visable, I dubble checked the coingecko ID, this could be correct. I remember it was working on Hydration and never on the AH part on Subscan. But lately both are not showing the price

* Update file size

---------

Co-authored-by: fabberlaan <205527636+lnrocket@users.noreply.github.com>